@@ -1,6 +1,6 @@
 {
   "name": "expenses-tracker",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "nx",
@@ -14,64 +14,64 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~15.2.0",
-    "@angular/cdk": "^15.2.8",
-    "@angular/common": "~15.2.0",
-    "@angular/compiler": "~15.2.0",
-    "@angular/core": "~15.2.0",
+    "@angular/animations": "16.0.6",
+    "@angular/cdk": "16.0.4",
+    "@angular/common": "16.0.6",
+    "@angular/compiler": "16.0.6",
+    "@angular/core": "16.0.6",
     "@angular/fire": "^7.5.0",
-    "@angular/forms": "~15.2.0",
-    "@angular/material": "^15.2.8",
-    "@angular/platform-browser": "~15.2.0",
-    "@angular/platform-browser-dynamic": "~15.2.0",
-    "@angular/router": "~15.2.0",
-    "@nrwl/angular": "15.9.2",
+    "@angular/forms": "16.0.6",
+    "@angular/material": "16.0.4",
+    "@angular/platform-browser": "16.0.6",
+    "@angular/platform-browser-dynamic": "16.0.6",
+    "@angular/router": "16.0.6",
+    "firebase": "^9.22.2",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "uuid": "^9.0.0",
-    "zone.js": "~0.12.0"
+    "zone.js": "0.13.1",
+    "@nx/angular": "16.3.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~15.2.0",
-    "@angular-devkit/core": "~15.2.0",
-    "@angular-devkit/schematics": "~15.2.0",
-    "@angular-eslint/eslint-plugin": "~15.0.0",
-    "@angular-eslint/eslint-plugin-template": "~15.0.0",
-    "@angular-eslint/template-parser": "~15.0.0",
-    "@angular/cli": "~15.2.0",
-    "@angular/compiler-cli": "~15.2.0",
-    "@angular/language-service": "~15.2.0",
-    "@nrwl/cypress": "15.9.2",
-    "@nrwl/eslint-plugin-nx": "15.9.2",
-    "@nrwl/jest": "15.9.2",
-    "@nrwl/js": "15.9.2",
-    "@nrwl/linter": "15.9.2",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/web": "15.9.2",
-    "@nrwl/workspace": "15.9.2",
-    "@schematics/angular": "~15.2.0",
+    "@angular-devkit/build-angular": "16.0.6",
+    "@angular-devkit/core": "16.0.6",
+    "@angular-devkit/schematics": "16.0.6",
+    "@angular-eslint/eslint-plugin": "16.0.3",
+    "@angular-eslint/eslint-plugin-template": "16.0.3",
+    "@angular-eslint/template-parser": "16.0.3",
+    "@angular/cli": "~16.0.0",
+    "@angular/compiler-cli": "16.0.6",
+    "@angular/language-service": "16.0.6",
+    "@schematics/angular": "16.0.6",
     "@types/jest": "^29.4.0",
     "@types/node": "16.11.7",
     "@types/uuid": "^9.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.36.1",
-    "@typescript-eslint/parser": "^5.36.1",
+    "@typescript-eslint/eslint-plugin": "5.60.0",
+    "@typescript-eslint/parser": "5.60.0",
     "autoprefixer": "^10.4.0",
-    "cypress": "^12.2.0",
+    "cypress": "12.15.0",
     "eslint": "~8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "eslint-plugin-prettier": "^4.2.1",
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
-    "jest-preset-angular": "~13.0.0",
-    "nx": "15.9.2",
+    "jest-preset-angular": "13.1.1",
+    "nx": "16.3.2",
     "postcss": "^8.4.5",
     "prettier": "^2.6.2",
     "source-map-explorer": "^2.5.3",
     "tailwindcss": "^3.0.2",
-    "ts-jest": "^29.0.5",
+    "ts-jest": "29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "~4.9.5"
+    "typescript": "5.0.4",
+    "nx-cloud": "16.0.5",
+    "@nx/workspace": "16.3.2",
+    "@nx/cypress": "16.3.2",
+    "@nx/js": "16.3.2",
+    "@nx/linter": "16.3.2",
+    "@nx/eslint-plugin": "16.3.2",
+    "@nx/jest": "16.3.2",
+    "@nx/web": "16.3.2"
   }
 }
-
