@@ -1,12 +1,12 @@
 import { Clipboard } from '@angular/cdk/clipboard';
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { FirebaseError } from '@angular/fire/app';
 import { Router } from '@angular/router';
 import { User } from 'firebase/auth';
 import { BehaviorSubject, catchError, of, tap, throwError } from 'rxjs';
 
-import { AuthService, ContextService, ErrorService } from '@expenses-tracker/core';
-import { NotificationService } from '@expenses-tracker/shared/common';
+import { AuthService, ContextService } from '@expenses-tracker/core';
+import { NotificationService, RoutePaths } from '@expenses-tracker/shared/common';
 import { IFlag, INITIAL_FLAGS } from '@expenses-tracker/shared/interfaces';
 
 export type ComponentFlags = {
@@ -27,14 +27,11 @@ export class ProfileService {
   #editMode$ = new BehaviorSubject<boolean>(false);
   #editMode = false;
 
-  constructor(
-    private _auth: AuthService,
-    private _context: ContextService,
-    private _error: ErrorService,
-    private _notification: NotificationService,
-    private _router: Router,
-    private _clipboard: Clipboard
-  ) {}
+  #auth = inject(AuthService);
+  #context = inject(ContextService);
+  #notification = inject(NotificationService);
+  #router = inject(Router);
+  #clipboard = inject(Clipboard);
 
   getUser$() {
     this.#flags = {
@@ -46,7 +43,7 @@ export class ProfileService {
       }
     };
     this.#setFlags(this.#flags);
-    return this._context.watchUser$().pipe(
+    return this.#context.watchUser$().pipe(
       tap(() => {
         this.#flags = {
           ...this.#flags,
@@ -85,17 +82,17 @@ export class ProfileService {
       }
     };
     this.#setFlags(this.#flags);
-    return this._auth.logout$().pipe(
+    return this.#auth.logout$().pipe(
       tap(() => {
-        this._notification.info({
+        this.#notification.info({
           title: 'Logout Successful',
           description: 'User has been logged out successfully'
         });
         this.#resetFlags();
-        this._router.navigate(['auth', 'login']);
+        this.#router.navigate([RoutePaths.Auth, RoutePaths.AuthLogin]);
       }),
       catchError(error => {
-        this._notification.error({
+        this.#notification.error({
           description: `${error}.`,
           title: 'Login failed'
         });
@@ -115,18 +112,18 @@ export class ProfileService {
   }
 
   updateUserInfo$(user: User) {
-    return this._auth.updateUserInfo$(user);
+    return this.#auth.updateUserInfo$(user);
   }
 
   copyUID(uid: string | null) {
-    const copyState = this._clipboard.copy(uid ?? '');
+    const copyState = this.#clipboard.copy(uid ?? '');
     if (copyState) {
-      this._notification.info({
+      this.#notification.info({
         title: 'Success!',
         description: 'UID copied to the clipboard'
       });
     } else {
-      this._notification.info({
+      this.#notification.info({
         title: 'Fail!',
         description: 'UID cannot be copied to the clipboard'
       });
