@@ -3,6 +3,8 @@ import { AngularFireAuth } from '@angular/fire/compat/auth';
 import { Router } from '@angular/router';
 import { map } from 'rxjs';
 
+import { RoutePaths } from '@expenses-tracker/shared/common';
+
 @Injectable({
   providedIn: 'root'
 })
@@ -16,7 +18,7 @@ export class AuthGuard {
         if (!!user) {
           return true;
         } else {
-          this.#router.navigate(['auth', 'login']);
+          this.#router.navigate([RoutePaths.Auth, RoutePaths.AuthLogin]);
           return false;
         }
       })
@@ -37,7 +39,7 @@ export class ReverseAuthGuard {
         if (!user) {
           return true;
         } else {
-          this.#router.navigate(['dashboard']);
+          this.#router.navigate([RoutePaths.Dashboard]);
           return false;
         }
       })
